packages/ocoi-converter/pyproject.toml
[project]
name = "ocoi-converter"
version = "0.1.0"
description = "PDF to Markdown conversion with Hebrew OCR support"
requires-python = ">=3.12"
dependencies = [
    "ocoi-common",
    "ocoi-db",
    "marker-pdf>=1.0",
    "pymupdf4llm>=0.0.10",
    "pymupdf>=1.24",
    "click>=8.0",
    "tqdm>=4.66",
]

[project.scripts]
ocoi-convert = "ocoi_converter.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocoi_converter"]

[tool.uv.sources]
ocoi-common = { workspace = true }
ocoi-db = { workspace = true }
